requirements.txt
altgraph==0.17.4
lief==0.16.3
numpy==2.2.3
opencv-python==4.11.0.86
packaging==24.2
pefile==2023.2.7
pillow==11.1.0
pillow_heif==0.21.0
pyinstaller==6.12.0
pyinstaller-hooks-contrib==2025.1
pywin32-ctypes==0.2.3
setuptools==75.8.2
tk==0.1.0
ttkbootstrap==1.10.1
